introduce AbstractParserTest
