+++ src/test/java/org/codeforamerica/shiba/application/parsers/HomeAddressParserTest.java
@@ -1,44 +1,21 @@
 package org.codeforamerica.shiba.application.parsers;
 
-import org.codeforamerica.shiba.YamlPropertySourceFactory;
 import org.codeforamerica.shiba.pages.data.ApplicationData;
 import org.codeforamerica.shiba.pages.data.InputData;
 import org.codeforamerica.shiba.pages.data.PageData;
 import org.codeforamerica.shiba.pages.data.PagesData;
 import org.codeforamerica.shiba.pages.enrichment.Address;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.test.context.ActiveProfiles;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.NONE;
 
-@SpringBootTest(webEnvironment = NONE)
-@ActiveProfiles("test")
-class HomeAddressParserTest {
-    @Autowired
-    HomeAddressParser homeAddressParser;
-
-    @TestConfiguration
-    @PropertySource(value = "classpath:test-parsing-config.yaml", factory = YamlPropertySourceFactory.class)
-    static class TestPageConfiguration {
-        @SuppressWarnings("ConfigurationProperties")
-        @Bean
-        @ConfigurationProperties(prefix = "test-parsing")
-        public ParsingConfiguration parsingConfiguration() {
-            return new ParsingConfiguration();
-        }
-    }
-
+class HomeAddressParserTest extends AbstractParserTest {
     @Test
     void shouldParseApplicationData() {
+        HomeAddressParser homeAddressParser = new HomeAddressParser(parsingConfiguration);
+
         ApplicationData applicationData = new ApplicationData();
         PagesData pagesData = new PagesData();
         PageData homePageData = new PageData();
